\name{trajSurvR-package}
\alias{trajSurvR-package}
\alias{trajSurvR}
\docType{package}
\title{Longitudinal Trajectory and Survival Analysis for Proteomic Data}
\description{
\pkg{trajSurvR} provides a collection of functions for analysing longitudinal
proteomic data, from quality control through differential expression,
correlation study, GAM/GAMM trajectory modelling, time-to-event analysis,
event prediction, onset-time estimation, and feature selection.

See \code{vignette("trajSurvR_demo", package = "trajSurvR")} for an
end-to-end worked example using simulated data.
}
\author{
\strong{Maintainer}: Ximing Ran \email{ranxm2@users.noreply.github.com}
}
\seealso{
\code{\link{simulate_trajsurv_data}},
\code{\link{qc_olink_data}},
\code{\link{fit_differential_mixed_models}},
\code{\link{run_trajectory_batch}},
\code{\link{fit_baseline_cox}}
}
\keyword{package}
